Move two public APIs to each optimizer rule file; Create a new file for shared static functions
