+++ arangod/Aql/Condition.cpp
@@ -26,7 +26,6 @@
 #include "Aql/Ast.h"
 #include "Aql/AstNode.h"
 #include "Aql/Collection.h"
-#include "Aql/ExecutionNode/CalculationNode.h"
 #include "Aql/ExecutionNode/EnumerateCollectionNode.h"
 #include "Aql/ExecutionPlan.h"
 #include "Aql/Expression.h"
@@ -45,7 +44,6 @@
 #include "Containers/FlatHashSet.h"
 #include "Containers/SmallVector.h"
 #include "Indexes/Index.h"
-#include "Logger/LogMacros.h"
 #include "Transaction/CountCache.h"
 #include "Transaction/Methods.h"
 
@@ -528,367 +526,6 @@ AstNode* Condition::createSimpleCondition(AstNode* node) const {
   orNode->addMember(andNode);
 
   return orNode;
-}
-
-namespace {
-bool areInRangeCallsIdentical(AstNode const* arrayNode,
-                              AstNode const* otherArrayNode,
-                              bool isFromTraverser) {
-  for (size_t i = 0; i < arrayNode->numMembers(); ++i) {
-    auto const* functionCallNodeElement = arrayNode->getMemberUnchecked(i);
-    auto const* functionCallNodeOtherElement =
-        otherArrayNode->getMemberUnchecked(i);
-
-    if (functionCallNodeElement->type != functionCallNodeOtherElement->type) {
-      return false;
-    }
-
-    switch (functionCallNodeElement->type) {
-      case NODE_TYPE_ATTRIBUTE_ACCESS: {
-        std::pair<Variable const*, std::vector<basics::AttributeName>>
-            attributeAccessForVariableResult;
-        std::pair<Variable const*, std::vector<basics::AttributeName>>
-            attributeAccessForVariableOtherResult;
-
-        if (!functionCallNodeElement->isAttributeAccessForVariable(
-                attributeAccessForVariableResult, isFromTraverser) ||
-            !functionCallNodeOtherElement->isAttributeAccessForVariable(
-                attributeAccessForVariableOtherResult, isFromTraverser) ||
-            attributeAccessForVariableResult.first !=
-                attributeAccessForVariableOtherResult.first ||
-            !basics::AttributeName::isIdentical(
-                attributeAccessForVariableResult.second,
-                attributeAccessForVariableOtherResult.second, false)) {
-          return false;
-        }
-        break;
-      }
-      case NODE_TYPE_VALUE: {
-        if (aql::compareAstNodes(functionCallNodeElement,
-                                 functionCallNodeOtherElement, true) != 0) {
-          return false;
-        }
-        break;
-      }
-      default: {
-        // We are being extremely pessimistic, meaning if we encounter
-        // anything else we will not remove the expression
-        return false;
-      }
-    }
-  }
-
-  return true;
-}
-
-bool canInRangeBeRemoved(AstNode const* inRangeNode,
-                         AstNode const* otherAndNode, bool isFromTraverser,
-                         Variable const* variable, Index const* index) {
-  for (std::size_t i{0}; i < otherAndNode->numMembers(); ++i) {
-    auto const* operand = otherAndNode->getMemberUnchecked(i);
-    if (operand->type != NODE_TYPE_FCALL ||
-        aql::functions::getFunctionName(*operand) != "IN_RANGE") {
-      continue;
-    }
-
-    // since we know that this is IN_RANGE node, it must contain 5 members
-    auto const* operandArrayNode = operand->getMember(0);
-    TRI_ASSERT(operandArrayNode->numMembers() == 5);
-
-    auto const* firstElemInRangeFunction = operandArrayNode->getMember(0);
-
-    bool isFieldCoveredByIndex{false};
-    std::pair<Variable const*, std::vector<basics::AttributeName>> result;
-    for (auto const& elem : index->fields()) {
-      if (firstElemInRangeFunction->type != NODE_TYPE_ATTRIBUTE_ACCESS ||
-          !firstElemInRangeFunction->isAttributeAccessForVariable(
-              result, isFromTraverser) ||
-          result.first != variable ||
-          !basics::AttributeName::isIdentical(result.second, elem, false)) {
-        clearAttributeAccess(result);
-        isFieldCoveredByIndex = true;
-        break;
-      }
-      clearAttributeAccess(result);
-    }
-
-    if (!isFieldCoveredByIndex) {
-      return false;
-    }
-
-    auto const* inRangeArrayNode = inRangeNode->getMember(0);
-    if (areInRangeCallsIdentical(inRangeArrayNode, operandArrayNode,
-                                 isFromTraverser)) {
-      return true;
-    }
-  }
-
-  return false;
-}
-}  // namespace
-
-bool Condition::isConditionCoveredBy(ExecutionPlan const* plan,
-                                     Variable const* variable,
-                                     AstNode const* condition,
-                                     AstNode const* otherAndNode) {
-  TRI_ASSERT(condition != nullptr);
-  TRI_ASSERT(otherAndNode != nullptr);
-  TRI_ASSERT(otherAndNode->type == NODE_TYPE_OPERATOR_NARY_AND);
-
-  if (!condition->isComparisonOperator() &&
-      !condition->isArrayComparisonOperator()) {
-    return false;
-  }
-
-  constexpr bool allowIndexedAccessInArray = true;
-  std::pair<Variable const*, std::vector<basics::AttributeName>> result;
-
-  auto lhs =
-      const_cast<AstNode*>(plan->resolveVariableAlias(condition->getMember(0)));
-  auto rhs =
-      const_cast<AstNode*>(plan->resolveVariableAlias(condition->getMember(1)));
-
-  if (lhs->type == NODE_TYPE_ATTRIBUTE_ACCESS ||
-      lhs->type == NODE_TYPE_EXPANSION) {
-    clearAttributeAccess(result);
-
-    if (lhs->isAttributeAccessForVariable(result, allowIndexedAccessInArray) &&
-        result.first == variable) {
-      ConditionPart current(variable, result.second, condition, ATTRIBUTE_LEFT,
-                            nullptr);
-
-      if (canRemove(plan, current, otherAndNode, allowIndexedAccessInArray)) {
-        return true;
-      }
-    }
-  }
-
-  if (rhs->type == NODE_TYPE_ATTRIBUTE_ACCESS ||
-      rhs->type == NODE_TYPE_EXPANSION) {
-    clearAttributeAccess(result);
-
-    if (rhs->isAttributeAccessForVariable(result, allowIndexedAccessInArray) &&
-        result.first == variable) {
-      ConditionPart current(variable, result.second, condition, ATTRIBUTE_RIGHT,
-                            nullptr);
-
-      if (canRemove(plan, current, otherAndNode, allowIndexedAccessInArray)) {
-        return true;
-      }
-    }
-  }
-
-  return false;
-}
-
-containers::HashSet<size_t> Condition::collectOverlappingMembersForIndex(
-    ExecutionPlan const* plan, Variable const* variable, AstNode const* andNode,
-    AstNode const* otherAndNode, Index const* index) {
-  TRI_ASSERT(index != nullptr);
-
-  containers::HashSet<size_t> toRemove;
-  bool const isSparse = index->sparse();
-  constexpr bool allowIndexedAccessInArray = true;
-
-  std::pair<Variable const*, std::vector<basics::AttributeName>> result;
-
-  for (size_t i = 0; i < andNode->numMembers(); ++i) {
-    auto operand = andNode->getMemberUnchecked(i);
-
-    if (isSparse && operand->isComparisonOperator() &&
-        (operand->type == NODE_TYPE_OPERATOR_BINARY_NE ||
-         operand->type == NODE_TYPE_OPERATOR_BINARY_GT)) {
-      // look for `!= null` and `> null`
-      // these can be removed if we are working with a sparse index!    
-      auto lhs = const_cast<AstNode*>(
-          plan->resolveVariableAlias(operand->getMemberUnchecked(0)));
-      auto rhs = const_cast<AstNode*>(
-          plan->resolveVariableAlias(operand->getMemberUnchecked(1)));
-
-      clearAttributeAccess(result);
-
-      if (rhs->isNullValue() &&
-          lhs->isAttributeAccessForVariable(result,
-                                            allowIndexedAccessInArray) &&
-          result.first == variable) {
-        auto const mayRemove = [&] {
-          if (auto ty = index->type();
-              ty == Index::TRI_IDX_TYPE_MDI_INDEX ||
-              ty == Index::TRI_IDX_TYPE_MDI_PREFIXED_INDEX) {
-            // For an MDI all fields are equal, and we are allowed to drop
-            // conditions for non-null on every attribute in the sparse case.    
-            return true;
-          }
-          // otherwise only remove the condition if the index is exactly on the
-          // same attribute as the condition
-
-          return index->fields().size() == 1 &&
-                 basics::AttributeName::isIdentical(result.second,
-                                                    index->fields()[0], false);
-        }();
-
-        if (mayRemove) {
-          toRemove.emplace(i);
-          continue;
-        }
-      }
-    }
-
-    if (operand->type == NODE_TYPE_FCALL &&
-        functions::getFunctionName(*operand) == "IN_RANGE") {
-      if (canInRangeBeRemoved(operand, otherAndNode,
-                              /*isFromTraverser*/ false, variable, index)) {
-        toRemove.emplace(i);
-      }
-      continue;
-    }
-
-    if (!operand->isComparisonOperator() ||
-        operand->type == NODE_TYPE_OPERATOR_BINARY_NE ||
-        operand->type == NODE_TYPE_OPERATOR_BINARY_NIN) {
-      continue;
-    }
-
-    if (isConditionCoveredBy(plan, variable, operand, otherAndNode)) {
-      toRemove.emplace(i);
-    }
-  }
-
-  return toRemove;
-}
-
-containers::HashSet<size_t> Condition::collectOverlappingMembersForTraversal(
-    ExecutionPlan const* plan, Variable const* variable, AstNode const* andNode,
-    AstNode const* otherAndNode, bool isPathCondition) {
-  containers::HashSet<size_t> toRemove;
-
-  for (size_t i = 0; i < andNode->numMembers(); ++i) {
-    auto operand = andNode->getMemberUnchecked(i);
-
-    if (operand->type == NODE_TYPE_FCALL &&
-        functions::getFunctionName(*operand) == "IN_RANGE") {
-      if (canInRangeBeRemoved(operand, otherAndNode,
-                              /*isFromTraverser*/ true, variable,
-                              /*index*/ nullptr)) {
-        toRemove.emplace(i);
-      }
-      continue;
-    }
-
-    bool allowOps = operand->isComparisonOperator();
-    if (isPathCondition) {
-      allowOps = allowOps || operand->isArrayComparisonOperator();
-    }
-
-    if (!allowOps) {
-      continue;
-    }
-
-    if (isConditionCoveredBy(plan, variable, operand, otherAndNode)) {
-      toRemove.emplace(i);
-    }
-  }
-
-  return toRemove;
-}
-
-bool Condition::extractSingleAndNodes(AstNode const* root,
-                                      AstNode const* condition,
-                                      AstNode const*& andNode,
-                                      AstNode const*& conditionAndNode) {
-  if (root == nullptr || condition == nullptr) {
-    return false;
-  }
-
-  TRI_ASSERT(root->type == NODE_TYPE_OPERATOR_NARY_OR);
-  TRI_ASSERT(condition->type == NODE_TYPE_OPERATOR_NARY_OR);
-
-  if (condition->numMembers() != 1 && root->numMembers() != 1) {
-    return false;
-  }
-
-  andNode = root->getMemberUnchecked(0);
-  conditionAndNode = condition->getMemberUnchecked(0);
-
-  TRI_ASSERT(andNode->type == NODE_TYPE_OPERATOR_NARY_AND);
-  TRI_ASSERT(conditionAndNode->type == NODE_TYPE_OPERATOR_NARY_AND);
-
-  return true;
-}
-
-AstNode* Condition::rebuildConditionWithoutMembers(
-    Ast* ast, AstNode const* andNode,
-    containers::HashSet<size_t> const& toRemove) {
-  if (toRemove.empty()) {
-    return const_cast<AstNode*>(andNode);
-  }
-
-  AstNode* newNode = nullptr;
-  for (size_t i = 0; i < andNode->numMembers(); ++i) {
-    if (toRemove.find(i) == toRemove.end()) {
-      auto what = andNode->getMemberUnchecked(i);
-      if (newNode == nullptr) {
-        // the only node so far
-        newNode = what;
-      } else {
-        // AND-combine with existing node
-        newNode = ast->createNodeBinaryOperator(NODE_TYPE_OPERATOR_BINARY_AND,
-                                                newNode, what);
-      }
-    }
-  }
-
-  return newNode;
-}
-
-/// @brief removes condition parts from another
-AstNode* Condition::removeIndexCondition(ExecutionPlan const* plan,
-                                         Variable const* variable,
-                                         AstNode const* condition,
-                                         Index const* index) {
-  TRI_ASSERT(index != nullptr);
-
-  if (_root == nullptr || condition == nullptr) {
-    return _root;
-  }
-
-  AstNode const* andNode = nullptr;
-  AstNode const* conditionAndNode = nullptr;
-  if (!extractSingleAndNodes(_root, condition, andNode, conditionAndNode)) {
-    return _root;
-  }
-
-  auto toRemove = collectOverlappingMembersForIndex(plan, variable, andNode,
-                                                    conditionAndNode, index);
-  if (toRemove.empty()) {
-    return _root;
-  }
-
-  return rebuildConditionWithoutMembers(_ast, andNode, toRemove);
-}
-
-/// @brief remove filter conditions already covered by the traversal
-AstNode* Condition::removeTraversalCondition(ExecutionPlan const* plan,
-                                             Variable const* variable,
-                                             AstNode* other,
-                                             bool isPathCondition) {
-  if (_root == nullptr || other == nullptr) {
-    return _root;
-  }
-
-  AstNode const* andNode = nullptr;
-  AstNode const* conditionAndNode = nullptr;
-  if (!extractSingleAndNodes(_root, other, andNode, conditionAndNode)) {
-    return _root;
-  }
-
-  auto toRemove = collectOverlappingMembersForTraversal(
-      plan, variable, andNode, conditionAndNode, isPathCondition);
-  if (toRemove.empty()) {
-    return _root;
-  }
-
-  return rebuildConditionWithoutMembers(_ast, andNode, toRemove);
 }
 
 /// @brief remove (now) invalid variables from the condition
@@ -1459,114 +1096,6 @@ void Condition::validateAst(AstNode const* node, int level) {
   }
 }
 #endif
-
-/// @brief checks if the current condition is covered by the other
-bool Condition::canRemove(ExecutionPlan const* plan, ConditionPart const& me,
-                          aql::AstNode const* andNode,
-                          bool allowArrayExpansion) {
-  TRI_ASSERT(andNode != nullptr);
-  TRI_ASSERT(andNode->type == NODE_TYPE_OPERATOR_NARY_AND);
-
-  std::pair<Variable const*, std::vector<basics::AttributeName>> result;
-
-  size_t const n = andNode->numMembers();
-
-  auto normalize = [&plan](AstNode const* node) -> std::string {
-    if (node->type == NODE_TYPE_REFERENCE) {
-      auto setter =
-          plan->getVarSetBy(static_cast<Variable const*>(node->getData())->id);
-      if (setter != nullptr &&
-          setter->getType() == ExecutionNode::CALCULATION) {
-        auto cn = ExecutionNode::castTo<CalculationNode const*>(setter);
-        // use expression node instead
-        node = cn->expression()->node();
-      }
-    }
-    // return string representation
-    return node->toString();
-  };
-
-  std::string temp;
-
-  try {
-    for (size_t i = 0; i < n; ++i) {
-      auto operand = andNode->getMemberUnchecked(i);
-
-      if (operand->isComparisonOperator() ||
-          (allowArrayExpansion && operand->isArrayComparisonOperator())) {
-        auto lhs = operand->getMember(0);
-        auto rhs = operand->getMember(1);
-
-        if (lhs->type == NODE_TYPE_ATTRIBUTE_ACCESS ||
-            (allowArrayExpansion && lhs->type == NODE_TYPE_EXPANSION)) {
-          clearAttributeAccess(result);
-
-          if (lhs->isAttributeAccessForVariable(result, allowArrayExpansion) &&
-              result.first == me.variable) {
-            temp.clear();
-            TRI_AttributeNamesToString(result.second, temp);
-            if (temp == me.attributeName) {
-              if (rhs->isConstant()) {
-                ConditionPart indexCondition(result.first, result.second,
-                                             operand, ATTRIBUTE_LEFT, nullptr);
-
-                if (me.isCoveredBy(indexCondition, false)) {
-                  return true;
-                }
-              }
-              // non-constant condition
-              else if (me.operatorType == operand->type &&
-                       normalize(me.valueNode) == normalize(rhs)) {
-                return true;
-              }
-            }
-          }
-        }
-
-        if (rhs->type == NODE_TYPE_ATTRIBUTE_ACCESS ||
-            rhs->type == NODE_TYPE_EXPANSION) {
-          clearAttributeAccess(result);
-
-          if (rhs->isAttributeAccessForVariable(result, allowArrayExpansion) &&
-              result.first == me.variable) {
-            temp.clear();
-            TRI_AttributeNamesToString(result.second, temp);
-            if (temp == me.attributeName) {
-              if (lhs->isConstant()) {
-                ConditionPart indexCondition(result.first, result.second,
-                                             operand, ATTRIBUTE_RIGHT, nullptr);
-
-                if (me.isCoveredBy(indexCondition, true)) {
-                  return true;
-                }
-              }
-              // non-constant condition
-              else {
-                auto opType = operand->type;
-                if (aql::Ast::isReversibleOperator(opType)) {
-                  opType = aql::Ast::reverseOperator(opType);
-                }
-                if (me.operatorType == opType &&
-                    normalize(me.valueNode) == normalize(lhs)) {
-                  return true;
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  } catch (std::exception const& ex) {
-    // simply ignore any errors (except trace-logging) and return false.
-    // this is not an error, but just means we cannot compare the two
-    // conditions for equality because there is no implemented way for
-    // comparing some of their components.
-    LOG_TOPIC("9f37b", TRACE, Logger::QUERIES)
-        << "caught exception in Condition::canRemove(): " << ex.what();
-  }
-
-  return false;
-}
 
 /// @brief deduplicate IN condition values (and sort them).
 /// will return either the unmodified original node or a copy.
